chore: refactor common actions
